+++ pkg/analyzer/test/src/diagnostics/test_all.dart
@@ -265,8 +265,16 @@ import 'field_initializing_formal_not_assignable_test.dart'
     as field_initializing_formal_not_assignable;
 import 'field_must_be_external_in_struct_test.dart'
     as field_must_be_external_in_struct;
+import 'final_class_extended_outside_of_library_test.dart'
+    as final_class_extended_outside_of_library;
+import 'final_class_implemented_outside_of_library_test.dart'
+    as final_class_implemented_outside_of_library;
 import 'final_initialized_in_declaration_and_constructor_test.dart'
     as final_initialized_in_declaration_and_constructor;
+import 'final_mixin_implemented_outside_of_library_test.dart'
+    as final_mixin_implemented_outside_of_library;
+import 'final_mixin_mixed_in_outside_of_library_test.dart'
+    as final_mixin_mixed_in_outside_of_library;
 import 'final_not_initialized_constructor_test.dart'
     as final_not_initialized_constructor;
 import 'final_not_initialized_test.dart' as final_not_initialized;
@@ -1032,6 +1040,10 @@ main() {
     field_initializer_redirecting_constructor.main();
     field_initializing_formal_not_assignable.main();
     field_must_be_external_in_struct.main();
+    final_class_extended_outside_of_library.main();
+    final_class_implemented_outside_of_library.main();
+    final_mixin_implemented_outside_of_library.main();
+    final_mixin_mixed_in_outside_of_library.main();
     final_not_initialized_constructor.main();
     final_not_initialized.main();
     for_in_of_invalid_element_type.main();