+++ pkg/compiler/lib/src/js_emitter/runtime_type_generator.dart
@@ -266,9 +266,9 @@ class _TypeContainedInOutputUnitVisitor
 
   @override
   bool visitRecordType(RecordType type, OutputUnit argument) {
-    // TODO(sra): The record shape is a bit like an implicit interface
-    // type. Does it have an OutputUnit?
-    throw UnimplementedError();
+    // The interface type that implements an allocated record is not needed to
+    // do subtyping.
+    return visitList(type.fields, argument);
   }
 
   @override