+++ tests/language/class_modifiers/final/final_mixin_implement_outside_error_test.dart
@@ -9,13 +9,13 @@
 import 'final_mixin_implement_lib.dart';
 
 abstract class AOutside implements FinalMixin {}
-// ^
-// [analyzer] unspecified
+//                                 ^^^^^^^^^^
+// [analyzer] COMPILE_TIME_ERROR.INVALID_USE_OF_TYPE_OUTSIDE_LIBRARY
 // [cfe] unspecified
 
 class BOutside implements FinalMixin {
-// ^
-// [analyzer] unspecified
+//                        ^^^^^^^^^^
+// [analyzer] COMPILE_TIME_ERROR.INVALID_USE_OF_TYPE_OUTSIDE_LIBRARY
 // [cfe] unspecified
   @override
   int foo = 1;