+++ pkg/vm_service/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 10.1.2
+- Fix bug where code would try to call `.toJson()` on `String`s.
+
+## 10.1.0
+- Update to version `4.1` of the spec.
+- Add optional `includeSubclasses` and `includeImplementers` parameters to
+  `getInstances`.
+
 ## 10.0.0
 - Update to version `4.0` of the spec.
 - Update for incorrectly documented types for `WeakReference`'s `target`,